Info / ideas / maps changes
Added more info and ideas pertaining to possible 'gets' we would want to
use that relates to schedules. Added more details to the maps.

All content in maps is used for displaying and providing POSSIBLE info
to show what we can do, it does not look the best but that is not the
point right now